+++ sync-for-payables/src/codat_sync_for_payables/models/shared/accountmappingoption.py
@@ -89,7 +89,7 @@ class AccountMappingOption(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -104,3 +104,9 @@ class AccountMappingOption(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    AccountMappingOption.model_rebuild()
+except NameError:
+    pass