+++ sync-for-payables/src/codat_sync_for_payables/models/shared/bankaccountmappingoption.py
@@ -118,7 +118,7 @@ class BankAccountMappingOption(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -133,3 +133,9 @@ class BankAccountMappingOption(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    BankAccountMappingOption.model_rebuild()
+except NameError:
+    pass