+++ sync-for-payables/src/codat_sync_for_payables/models/shared/supplier.py
@@ -127,7 +127,7 @@ class Supplier(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -142,3 +142,9 @@ class Supplier(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    Supplier.model_rebuild()
+except NameError:
+    pass